Moved some common unit tests to the shared projects
